nlattr: add a special case to decode_nlattr
As of now, it's impossible to introduce a dispatching nlattr decoder
that performs actions based on type of the message.

This is going to be used later in IFLA_AF_SPEC decoding.

* nlattr.c (decode_nlattr_with_data): Handle zero size and non-NULL
decoders in a special way.
(decode_nlattr): Add an error message about ignoring of opaque_data
argument when zero size and non-NULL decoders are provided.
* nlattr.h (decode_nlattr): Add a comment regarding new special case.